Move entries to owned
